renamed helpers to utils for consistency
